Realtime-Lane-Segmentation/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.19.0

# Optional: TensorRT acceleration (requires CUDA)
# tensorrt>=8.0.0
# pycuda>=2021.1

# Additional dependencies for enhanced functionality
tqdm>=4.60.0 